FIX: 관리자, 사용자 CODE 병합
