+++ src/main/java/state/admin/userManage/application/processor/UserUpdateProcessor.java
@@ -3,7 +3,7 @@ package state.admin.userManage.application.processor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 import state.admin.userManage.application.command.UserInfoUpdateCommand;
-import state.admin.userManage.application.common.error.ErrorCode;
+import state.common.exception.ErrorCode;
 import state.admin.userManage.application.common.exception.ApiException;
 import state.admin.userManage.domain.entity.User;
 import state.admin.userManage.domain.repository.UserRepository;
