+++ src/main/java/state/common/exception/ErrorCode.java
@@ -1,20 +1,23 @@
 package state.common.exception;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@AllArgsConstructor
 public enum ErrorCode {
-    USER_NOT_FOUND("사용자를 찾을 수 없습니다."),
-    DEPARTMENT_NOT_EXIST("부서 정보가 존재하지 않습니다."),
-    DUPLICATE_USERID("이미 존재하는 아이디입니다."),
-    DUPLICATE_USER("이미 존재하는 이메일입니다."),
-    DEPARTMENT_NOT_FOUND("부서 정보를 찾을 수 없습니다."),
-    POSITION_NOT_FOUND("직위 정보를 찾을 수 없습니다.");
+    OK(200,200,"성공"),
 
-    private final String message;
+    BAD_REQUEST(HttpStatus.BAD_REQUEST.value(), 400, "잘못된 요청 입니다."),
+    ID_CHECK(HttpStatus.BAD_REQUEST.value(), 409, "중복된 아이디 입니다."),
 
-    ErrorCode(String message) {
-        this.message = message;
-    }
+    SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR.value(), 500, "서버에러"),
+    NULL_POINT(HttpStatus.INTERNAL_SERVER_ERROR.value(), 512, "Null Point"),
 
-    public String getMessage() {
-        return message;
-    }
+    USER_NOT_FOUND(HttpStatus.NOT_FOUND.value(), 404, "사용자를 찾을 수 없습니다.");
+
+    private final Integer httpStatusCode;
+    private final Integer errorCode;
+    private final String description;
 }
