+++ src/main/java/state/member/domain/exception/DuplicateEmailException.java
@@ -1,9 +1,8 @@
 package state.member.domain.exception;
 
-import state.common.exception.ErrorCode;
 
 public class DuplicateEmailException extends RuntimeException {
     public DuplicateEmailException() {
-        super(ErrorCode.DUPLICATE_USER.getMessage());
+        super(MemberErrorCode.DUPLICATE_USER.getDescription());
     }
 }
