+++ src/main/java/state/member/domain/exception/DuplicateUserIdException.java
@@ -1,9 +1,7 @@
 package state.member.domain.exception;
 
-import state.common.exception.ErrorCode;
-
 public class DuplicateUserIdException extends RuntimeException {
     public DuplicateUserIdException() {
-        super(ErrorCode.DUPLICATE_USERID.getMessage());
+        super(MemberErrorCode.DUPLICATE_USERID.getDescription());
     }
 }
