+++ src/main/java/state/member/domain/exception/MemberNotFoundException.java
@@ -4,6 +4,6 @@ import state.common.exception.ErrorCode;
 
 public class MemberNotFoundException extends RuntimeException {
     public MemberNotFoundException() {
-        super(ErrorCode.USER_NOT_FOUND.getMessage());
+        super(ErrorCode.USER_NOT_FOUND.getDescription());
     }
 }
