--- src/main/java/state/member/domain/exception/PositionNotExistException.java
@@ -1,9 +1,0 @@
-package state.member.domain.exception;
-
-import state.common.exception.ErrorCode;
-
-public class PositionNotExistException extends RuntimeException {
-    public PositionNotExistException() {
-        super(ErrorCode.POSITION_NOT_FOUND.getMessage());
-    }
-}
